requirements.txt
tensorflow>=2.8.0
mediapipe>=0.8.9
opencv-python>=4.5.4
numpy>=1.20.0
pygame>=2.0.1
pillow>=8.3.1
matplotlib>=3.5.0
requests>=2.26.0
tqdm>=4.62.3
tk>=0.1.0
imageio>=2.13.5 